requirements.txt
oauth2==1.9.0.post1
python-twitter==3.3
